preliminary clean up
