+++ .gitignore
@@ -1,6 +1,6 @@
 obj/
 dep/
-tmp/
+generated/
 junk/
 out_scm/
 tmp_scm/
