+++ Makefile
@@ -18,11 +18,11 @@ LAXCC=cc -Werror $(CCMODE)
 PYTHON=python3
 
 EXTRA_YAML=$(wildcard src/*.yaml)
-EXTRA_C_TARGETS=$(patsubst src/%.yaml,tmp/%.c,$(EXTRA_YAML))
-EXTRA_H_TARGETS=$(patsubst src/%.yaml,tmp/%.h,$(EXTRA_YAML))
-EXTRA_OBJTYPES_H_TARGETS=$(patsubst src/%.yaml,tmp/%_objtypes.h,$(EXTRA_YAML))
-EXTRA_DEBUG_H_TARGETS=$(patsubst src/%.yaml,tmp/%_debug.h,$(EXTRA_YAML))
-EXTRA_DEBUG_C_TARGETS=$(patsubst src/%.yaml,tmp/%_debug.c,$(EXTRA_YAML))
+EXTRA_C_TARGETS=$(patsubst src/%.yaml,generated/%.c,$(EXTRA_YAML))
+EXTRA_H_TARGETS=$(patsubst src/%.yaml,generated/%.h,$(EXTRA_YAML))
+EXTRA_OBJTYPES_H_TARGETS=$(patsubst src/%.yaml,generated/%_objtypes.h,$(EXTRA_YAML))
+EXTRA_DEBUG_H_TARGETS=$(patsubst src/%.yaml,generated/%_debug.h,$(EXTRA_YAML))
+EXTRA_DEBUG_C_TARGETS=$(patsubst src/%.yaml,generated/%_debug.c,$(EXTRA_YAML))
 
 EXTRA_TARGETS= \
     $(EXTRA_C_TARGETS) \
@@ -33,7 +33,7 @@ EXTRA_TARGETS= \
 
 MAIN=src/main.c
 CFILES=$(filter-out $(MAIN), $(wildcard src/*.c))
-EXTRA_CFILES=tmp/lexer.c tmp/parser.c $(EXTRA_C_TARGETS) $(EXTRA_DEBUG_C_TARGETS)
+EXTRA_CFILES=generated/lexer.c generated/parser.c $(EXTRA_C_TARGETS) $(EXTRA_DEBUG_C_TARGETS)
 TEST_CFILES=$(wildcard tests/src/*.c)
 
 TEST_TARGETS=$(patsubst tests/src/%.c,tests/%,$(TEST_CFILES))
@@ -48,14 +48,14 @@ MAIN_DEP=dep/main.d
 DEP=$(patsubst obj/%,dep/%,$(patsubst %.o,%.d,$(OBJ)))
 TEST_DEP=$(patsubst obj/%,dep/%,$(patsubst %.o,%.d,$(TEST_OBJ)))
 
-EXTRA_OBJ=$(patsubst tmp/%,obj/%,$(patsubst %.c,%.o,$(EXTRA_CFILES)))
+EXTRA_OBJ=$(patsubst generated/%,obj/%,$(patsubst %.c,%.o,$(EXTRA_CFILES)))
 EXTRA_DEP=$(patsubst obj/%,dep/%,$(patsubst %.o,%.d,$(EXTRA_OBJ)))
 
 ALL_OBJ=$(OBJ) $(EXTRA_OBJ)
 ALL_DEP=$(DEP) $(EXTRA_DEP) $(TEST_DEP)
 
-TMP_H=tmp/parser.h tmp/lexer.h
-TMP_C=tmp/parser.c tmp/lexer.c
+TMP_H=generated/parser.h generated/lexer.h
+TMP_C=generated/parser.c generated/lexer.c
 
 all: $(TARGET)
 
@@ -64,19 +64,19 @@ $(TARGET): $(MAIN_OBJ) $(ALL_OBJ)
 
 include $(ALL_DEP)
 
-$(EXTRA_C_TARGETS): tmp/%.c: src/%.yaml tools/makeAST.py | tmp
+$(EXTRA_C_TARGETS): generated/%.c: src/%.yaml tools/makeAST.py | generated
 	$(PYTHON) tools/makeAST.py $< c > $@ || (rm -f $@ ; exit 1)
 
-$(EXTRA_H_TARGETS): tmp/%.h: src/%.yaml tools/makeAST.py | tmp
+$(EXTRA_H_TARGETS): generated/%.h: src/%.yaml tools/makeAST.py | generated
 	$(PYTHON) tools/makeAST.py $< h > $@ || (rm -f $@ ; exit 1)
 
-$(EXTRA_OBJTYPES_H_TARGETS): tmp/%_objtypes.h: src/%.yaml tools/makeAST.py | tmp
+$(EXTRA_OBJTYPES_H_TARGETS): generated/%_objtypes.h: src/%.yaml tools/makeAST.py | generated
 	$(PYTHON) tools/makeAST.py $< objtypes_h > $@ || (rm -f $@ ; exit 1)
 
-$(EXTRA_DEBUG_H_TARGETS): tmp/%_debug.h: src/%.yaml tools/makeAST.py | tmp
+$(EXTRA_DEBUG_H_TARGETS): generated/%_debug.h: src/%.yaml tools/makeAST.py | generated
 	$(PYTHON) tools/makeAST.py $< debug_h > $@ || (rm -f $@ ; exit 1)
 
-$(EXTRA_DEBUG_C_TARGETS): tmp/%_debug.c: src/%.yaml tools/makeAST.py | tmp
+$(EXTRA_DEBUG_C_TARGETS): generated/%_debug.c: src/%.yaml tools/makeAST.py | generated
 	$(PYTHON) tools/makeAST.py $< debug_c > $@ || (rm -f $@ ; exit 1)
 
 
@@ -87,28 +87,28 @@ tags: src/* $(EXTRA_TARGETS) $(TMP_H) $(TMP_C)
 	ctags src/* $(EXTRA_TARGETS) $(TMP_H) $(TMP_C)
 
 $(MAIN_OBJ) $(OBJ): obj/%.o: src/%.c | obj
-	$(CC) -I tmp/ -I src/ -c $< -o $@
+	$(CC) -I generated/ -I src/ -c $< -o $@
 
-$(EXTRA_OBJ): obj/%.o: tmp/%.c | obj
-	$(LAXCC) -I src/ -I tmp/ -c $< -o $@
+$(EXTRA_OBJ): obj/%.o: generated/%.c | obj
+	$(LAXCC) -I src/ -I generated/ -c $< -o $@
 
 $(TEST_OBJ): obj/%.o: tests/src/%.c | obj
-	$(LAXCC) -I src/ -I tmp/ -c $< -o $@
+	$(LAXCC) -I src/ -I generated/ -c $< -o $@
 
 $(MAIN_DEP) $(DEP): dep/%.d: src/%.c .generated | dep
-	$(CC) -I tmp/ -I src/ -MM -MT $(patsubst dep/%,obj/%,$(patsubst %.d,%.o,$@)) -o $@ $<
+	$(CC) -I generated/ -I src/ -MM -MT $(patsubst dep/%,obj/%,$(patsubst %.d,%.o,$@)) -o $@ $<
 
-$(EXTRA_DEP): dep/%.d: tmp/%.c .generated | dep
-	$(LAXCC) -I src/ -I tmp/ -MM -MT $(patsubst dep/%,obj/%,$(patsubst %.d,%.o,$@)) -o $@ $<
+$(EXTRA_DEP): dep/%.d: generated/%.c .generated | dep
+	$(LAXCC) -I src/ -I generated/ -MM -MT $(patsubst dep/%,obj/%,$(patsubst %.d,%.o,$@)) -o $@ $<
 
 $(TEST_DEP): dep/%.d: tests/src/%.c .generated | dep
-	$(CC) -I src/ -I tmp/ -MM -MT $(patsubst dep/%,obj/%,$(patsubst %.d,%.o,$@)) -o $@ $<
+	$(CC) -I src/ -I generated/ -MM -MT $(patsubst dep/%,obj/%,$(patsubst %.d,%.o,$@)) -o $@ $<
 
-tmp/lexer.c tmp/lexer.h: src/lexer.l | tmp
-	flex --header-file=tmp/lexer.h -o tmp/lexer.c $<
+generated/lexer.c generated/lexer.h: src/lexer.l | generated
+	flex --header-file=generated/lexer.h -o generated/lexer.c $<
 
-tmp/parser.c tmp/parser.h: src/parser.y | tmp
-	bison -v -Werror --header=tmp/parser.h -o tmp/parser.c $<
+generated/parser.c generated/parser.h: src/parser.y | generated
+	bison -v -Werror --header=generated/parser.h -o generated/parser.c $<
 
 test: $(TEST_TARGETS)
 	for t in $(TEST_TARGETS) ; do $$t || exit 1 ; done
@@ -116,11 +116,11 @@ test: $(TEST_TARGETS)
 $(TEST_TARGETS): tests/%: obj/%.o $(ALL_OBJ)
 	$(CC) -o $@ $< $(ALL_OBJ) -lm
 
-dep obj tmp:
+dep obj generated:
 	mkdir $@
 
 clean: deps
-	rm -rf $(TARGET) obj callgrind.out.* tmp $(TEST_TARGETS)
+	rm -rf $(TARGET) obj callgrind.out.* generated $(TEST_TARGETS)
 
 deps:
 	rm -rf dep
