+++ src/debugging_on.h
@@ -22,18 +22,37 @@ static int _debugInvocationId __attribute__((unused)) = 0;
 static bool _debuggingOn __attribute__((unused)) = true;
 static int _debuggingDepth __attribute__((unused)) = 0;
 
+#define __DEBUGPAD__() do { for (int pads = _debuggingDepth / 4; pads > 0; pads--) { eprintf("   |"); } eprintf("%*s", _debuggingDepth % 4, ""); } while (false)
+
 #define DEBUG(...) do { \
     if (_debuggingOn) { \
-        eprintf("%s:%-5d %*s", __FILE__, __LINE__, _debuggingDepth, ""); \
+        eprintf("%s:%-5d", __FILE__, __LINE__); \
+        __DEBUGPAD__(); \
         eprintf(__VA_ARGS__); \
         eprintf("\n"); \
     } \
 } while(0)
+
+#define DEBUGN(...) do { \
+    if (_debuggingOn) { \
+        eprintf("%s:%-5d", __FILE__, __LINE__); \
+        __DEBUGPAD__(); \
+        eprintf(__VA_ARGS__); \
+    } \
+} while(0)
+
 #define ENTER(name) int _debugMyId = _debugInvocationId++; DEBUG("ENTER " #name " #%d", _debugMyId); _debuggingDepth++
+
 #define LEAVE(name) _debuggingDepth--; DEBUG("LEAVE " #name " #%d", _debugMyId)
+
 #define NEWLINE() do { if (_debuggingOn) eprintf("\n"); } while(0)
-#define IFDEBUG(x) do { if (_debuggingOn) { eprintf("%s:%-5d %*s", __FILE__, __LINE__, _debuggingDepth, ""); x; NEWLINE(); } } while(0)
+
+#define IFDEBUG(x) do { if (_debuggingOn) { eprintf("%s:%-5d", __FILE__, __LINE__); __DEBUGPAD__(); x; NEWLINE(); } } while(0)
+
+#define IFDEBUGN(x) do { if (_debuggingOn) { x; NEWLINE(); } } while(0)
+
 #define DEBUGGING_ON() do { _debuggingOn = true; } while (0)
+
 #define DEBUGGING_OFF() do { _debuggingOn = false; } while (0)
 
 #endif
