+++ src/lambda_conversion.h
@@ -22,5 +22,4 @@
 #include "lambda.h"
 
 LamExp *lamConvertNest(AstNest *nest, LamContext *env);
-LamExp *lamPerformSubstitutions(LamExp *exp, HashTable *substitutions);
 #endif
