+++ src/print_generator.c
@@ -14,12 +14,16 @@
  *
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ *
+ * print function generator
+ * invoked during lambda conversion
+ * generates a print function for each typedef
  */
 
 // print function generator, compiler and run-time code
 
 #include <stdio.h>
-#include "print.h"
+#include "print_generator.h"
 #include "cekf.h"
 #include "common.h"
 #include "lambda.h"
@@ -27,22 +31,13 @@
 #include "symbol.h"
 #include "symbols.h"
 
-#ifdef DEBUG_PRINT
+#ifdef DEBUG_PRINT_GENERATOR
 #include "debugging_on.h"
 #else
 #include "debugging_off.h"
 #endif
 
-static void putVec(Vec *x);
-
 static LamLetRecBindings *makePrintFunction(LamTypeDef *typeDef, LamLetRecBindings *next, LamContext *env, bool inPreamble);
-static HashSymbol *makePrintName(char *prefix, char *name);
-
-/****************************************************************
- * print function generator
- * invoked during lambda conversion
- * generates a print function for each typedef
- ****************************************************************/
 
 LamLetRecBindings *makePrintFunctions(LamTypeDefList *typeDefs, LamLetRecBindings *next, LamContext *env, bool inPreamble) {
     ENTER(makePrintFunctions);
@@ -64,7 +59,7 @@ static bool isListType(LamType *type) {
     return type->name == listSymbol();
 }
 
-static HashSymbol *makePrintName(char *prefix, char *name) {
+HashSymbol *makePrintName(char *prefix, char *name) {
     HashSymbol *res = NULL;
     int size = strlen(name) + strlen(prefix) + 1;
     if (size < 128) {
@@ -89,7 +84,7 @@ static int countLamVarList(LamVarList *list) {
     return res;
 }
 
-static int countLamList(LamList *list) {
+int countLamList(LamList *list) {
     int res = 0;
     while (list != NULL) {
         list = list->next;
@@ -197,17 +192,17 @@ static LamExp *makePrintAccessor(int index, LamTypeConstructorInfo *info) {
     return res;
 }
 
-static LamExp *makeSymbolExpr(char *name) {
+LamExp *makeSymbolExpr(char *name) {
     HashSymbol *symbol = newSymbol(name);
     LamExp *exp = newLamExp(LAMEXP_TYPE_VAR, LAMEXP_VAL_VAR(symbol));
     return exp;
 }
 
-static LamExp *makePrintInt() {
+LamExp *makePrintInt() {
     return makeSymbolExpr("print$int");
 }
 
-static LamExp *makePrintChar() {
+LamExp *makePrintChar() {
     return makeSymbolExpr("print$char");
 }
 
@@ -438,164 +433,4 @@ static LamLetRecBindings *makePrintFunction(LamTypeDef *typeDef, LamLetRecBindin
     } else {
         return makePrintTypeFunction(typeDef, env, next);
     }
-}
-
-/****************************************************************
- * print function compiler
- * invoked during type checking
- * computes a print function that will print the given type
- ****************************************************************/
-static LamExp *compilePrinterForFunction(TcFunction *function);
-static LamExp *compilePrinterForPair(TcPair *pair);
-static LamExp *compilePrinterForVar(TcVar *var, TcEnv *env);
-static LamExp *compilePrinterForInt();
-static LamExp *compilePrinterForChar();
-static LamExp *compilePrinterForTypeDef(TcTypeDef *typeDef, TcEnv *env);
-
-LamExp *compilePrinterForType(TcType *type, TcEnv *env) {
-    LamExp *res = NULL;
-    switch (type->type) {
-        case TCTYPE_TYPE_FUNCTION:
-            res = compilePrinterForFunction(type->val.function);
-            break;
-        case TCTYPE_TYPE_PAIR:
-            res = compilePrinterForPair(type->val.pair);
-            break;
-        case TCTYPE_TYPE_VAR:
-            res = compilePrinterForVar(type->val.var, env);
-            break;
-        case TCTYPE_TYPE_SMALLINTEGER:
-        case TCTYPE_TYPE_BIGINTEGER:
-            res = compilePrinterForInt();
-            break;
-        case TCTYPE_TYPE_CHARACTER:
-            res = compilePrinterForChar();
-            break;
-        case TCTYPE_TYPE_TYPEDEF:
-            res = compilePrinterForTypeDef(type->val.typeDef, env);
-            break;
-        default:
-            cant_happen("unrecognised TcType %d in compilePrinterForType", type->type);
-    }
-    return res;
-}
-
-static LamExp *compilePrinterForFunction(TcFunction *function __attribute__((unused))) {
-    return makeSymbolExpr("print$fn");
-}
-
-static LamExp *compilePrinterForPair(TcPair *pair __attribute__((unused))) {
-    cant_happen("compilePrinterForPair not implemented yet");
-}
-
-static LamExp *compilePrinterForVar(TcVar *var, TcEnv *env) {
-    if (var->instance == NULL) {
-        return makeSymbolExpr("print$");
-    }
-    return compilePrinterForType(var->instance, env);
-}
-
-static LamExp *compilePrinterForInt() {
-    return makePrintInt();
-}
-
-static LamExp *compilePrinterForChar() {
-    return makePrintChar();
-}
-
-static LamList *compilePrinterForTypeDefArgs(TcTypeDefArgs *args, TcEnv *env) {
-    if (args == NULL) return NULL;
-    LamList *next = compilePrinterForTypeDefArgs(args->next, env);
-    int save = PROTECT(next);
-    LamExp *this = compilePrinterForType(args->type, env);
-    PROTECT(this);
-    LamList *res = newLamList(this, next);
-    UNPROTECT(save);
-    return res;
-}
-
-static LamExp *compilePrinterForString() {
-    HashSymbol *name = newSymbol("print$string");
-    return newLamExp(LAMEXP_TYPE_VAR, LAMEXP_VAL_VAR(name));
-}
-
-static LamExp *compilePrinterForTypeDef(TcTypeDef *typeDef, TcEnv *env) {
-    if (typeDef->name == listSymbol()) {
-        if (typeDef->args && typeDef->args->type->type == TCTYPE_TYPE_CHARACTER) {
-            return compilePrinterForString();
-        }
-    }
-    HashSymbol *name = makePrintName("print$", typeDef->name->name);
-    if (!getFromTcEnv(env, name, NULL)) {
-        return makeSymbolExpr("print$");
-    }
-    LamExp *exp = newLamExp(LAMEXP_TYPE_VAR, LAMEXP_VAL_VAR(name));
-    int save = PROTECT(exp);
-    LamList *args = compilePrinterForTypeDefArgs(typeDef->args, env);
-    PROTECT(args);
-    int nargs = countLamList(args);
-    if (nargs == 0) {
-        UNPROTECT(save);
-        return exp;
-    }
-    LamApply *apply = newLamApply(exp, nargs, args);
-    PROTECT(apply);
-    LamExp *res = newLamExp(LAMEXP_TYPE_APPLY, LAMEXP_VAL_APPLY(apply));
-    UNPROTECT(save);
-    return res;
-}
-
-
-/****************************************************************
- * run-time print code supporting putv
- * used when the type cannot be detrmined.
- ****************************************************************/
-
-void putValue(Value x) {
-    switch (x.type) {
-        case VALUE_TYPE_VOID:
-            printf("<void>");
-            break;
-        case VALUE_TYPE_STDINT:
-            printf("%d", x.val.z);
-            break;
-        case VALUE_TYPE_BIGINT:
-            fprintBigInt(stdout, x.val.b);
-            break;
-        case VALUE_TYPE_CHARACTER:
-            switch (x.val.c) {
-                case '\t':
-                    printf("'\\t'");
-                    break;
-                case '\n':
-                    printf("'\\n'");
-                    break;
-                default:
-                    printf("'%c'", x.val.c);
-                    break;
-            }
-            break;
-        case VALUE_TYPE_CLO:
-            printf("<closure>");
-            break;
-        case VALUE_TYPE_CONT:
-            printf("<continuation>");
-            break;
-        case VALUE_TYPE_VEC:
-            putVec(x.val.vec);
-            break;
-        default:
-            cant_happen("unrecognised value type in putValue");
-    }
-}
-
-static void putVec(Vec *x) {
-    printf("#[");
-    for (int i = 0; i < x->size; i++) {
-        putValue(x->values[i]);
-        if (i + 1 < x->size) {
-            printf(" ");
-        }
-    }
-    printf("]");
 }
