+++ src/print_generator.h
@@ -1,5 +1,5 @@
-#ifndef cekf_debugging
-#define cekf_debugging
+#ifndef cekf_print_generator_h
+#define cekf_print_generator_h
 /*
  * CEKF - VM supporting amb
  * Copyright (C) 2022-2023  Bill Hails
@@ -16,18 +16,17 @@
  *
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
- *
- * Term Pattern Matching Compiler stage 4. code generation
  */
 
-#define ENTER(n)
-#define LEAVE(n)
-#define DEBUG(...)
-#define DEBUGN(...)
-#define IFDEBUG(x)
-#define IFDEBUGN(x)
-#define NEWLINE()
-#define DEBUGGING_ON()
-#define DEBUGGING_OFF()
+#include "lambda.h"
+#include "value.h"
+#include "tc.h"
 
-#endif
+LamLetRecBindings *makePrintFunctions(LamTypeDefList *typeDefs, LamLetRecBindings *rest, LamContext *env, bool inPreamble);
+LamExp *makeSymbolExpr(char *name);
+LamExp *makePrintInt();
+LamExp *makePrintChar();
+HashSymbol *makePrintName(char *prefix, char *name);
+int countLamList(LamList *list);
+
+ #endif
